+++ CHANGELOG.md
@@ -1309,3 +1309,4 @@
 - [2026-05-20 21:31] fix(storage): fix memory leak by cancelling storage subscription
 - [2026-05-20 21:32] fix(navigation): fix memory leak by cancelling navigation subscription
 - [2026-05-20 22:06] feat(core): implement responsive core view component
+- [2026-05-22 10:32] chore(test): update dependencies and cleanup unused imports in test